CodingDojo/games/moebius/src/main/java/com/codenjoy/dojo/moebius/services/Scores.java
package com.codenjoy.dojo.moebius.services;

import com.codenjoy.dojo.services.PlayerScores;
import com.codenjoy.dojo.services.settings.Parameter;
import com.codenjoy.dojo.services.settings.Settings;

public class Scores implements PlayerScores {

    private final Parameter<Integer> winScore;
    private final Parameter<Integer> loosePenalty;

    private volatile int score;

    public Scores(int startScore, Settings settings) {
        this.score = startScore;

        winScore = settings.addEditBox("Win score").type(Integer.class).def(30);
        loosePenalty = settings.addEditBox("Loose penalty").type(Integer.class).def(100);
    }

    @Override
    public int clear() {
        return score = 0;
    }

    @Override
    public int getScore() {
        return score;
    }

    @Override
    public void event(Object o) {
        switch (((Events) o).getType()) {
            case WIN:
                score += winScore.getValue();
                break;
            case GAME_OVER:
                score -= loosePenalty.getValue();
                break;
        }
        score = Math.max(0, score);
    }
}
